fix(design): apply design
