+++ src/components/PropertiesPanel/Properties/Properties.module.css
@@ -1,0 +1,33 @@
+h2 {
+  font-size: 16px;
+}
+
+label {
+  font-weight: 600;
+  font-size: 12px;
+}
+
+.Container {
+  flex-shrink: 0;
+  border-bottom: 1px solid #888888;
+  padding-bottom: 20px;
+  display: flex;
+  flex-direction: column;
+  gap: 10px;
+}
+
+.InputForm {
+  flex: 1 1 0;
+  display: flex;
+  flex-direction: column;
+  gap: 2.5px;
+}
+
+.FormContainer {
+  display: flex;
+  gap: 5px;
+}
+
+.Input {
+  width: 100%;
+}
